% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dailyfiles.R
\name{mod_dailyfiles_ui}
\alias{mod_dailyfiles_ui}
\alias{mod_dailyfiles_server}
\title{mod_dailyfiles_ui and mod_dailyfiles_server}
\usage{
mod_dailyfiles_ui(id)

mod_dailyfiles_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
