% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dailyfiles.R, R/mod_dailyfiles_trend.R
\name{mod_dailyfiles_trend_ui}
\alias{mod_dailyfiles_trend_ui}
\alias{mod_new_cases_server}
\alias{mod_dailyfiles_trend_server}
\title{mod_dailyfiles_trend_ui and mod_dailyfiles_trend_server}
\usage{
mod_dailyfiles_trend_ui(id)

mod_new_cases_server(input, output, session, df)

mod_dailyfiles_trend_ui(id)

mod_dailyfiles_trend_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.

A shiny Module.
}
\keyword{internal}
