% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R, R/mod_table_normal.R
\name{mod_table_ui}
\alias{mod_table_ui}
\alias{mod_table_server}
\alias{mod_table_normal_ui}
\alias{mod_table_normal_server}
\title{mod_table_ui and mod_table_server}
\usage{
mod_table_ui(id, label)

mod_table_server(input, output, session, df)

mod_table_normal_ui(id, label)

mod_table_normal_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.

A shiny Module.
}
\keyword{internal}
